Move migration code out of module. Fix exp.Material conditions, adding check on exp.ExperimentRun.
