+++ experiment/src/org/labkey/experiment/api/DataClassDomainKind.java
@@ -93,12 +93,13 @@ public class DataClassDomainKind extends AbstractDomainKind<DataClassDomainKindP
 
     private static final Set<String> FORCE_ENABLED_SYSTEM_FIELDS;
 
-    static {
+    static
+    {
         BASE_PROPERTIES = Collections.unmodifiableSet(Sets.newLinkedHashSet(Arrays.asList(
-                new PropertyStorageSpec("genId", JdbcType.INTEGER),
-                new PropertyStorageSpec("lsid", JdbcType.VARCHAR, 300).setNullable(false),
-                new PropertyStorageSpec("name", JdbcType.VARCHAR, 200),
-                new PropertyStorageSpec("classid", JdbcType.INTEGER)
+            new PropertyStorageSpec("genId", JdbcType.INTEGER),
+            new PropertyStorageSpec("lsid", JdbcType.VARCHAR, 300).setNullable(false),
+            new PropertyStorageSpec("name", JdbcType.VARCHAR, 200),
+            new PropertyStorageSpec("classid", JdbcType.INTEGER)
         )));
 
         Set<String> names = new HashSet<>();
